+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-## 1.0.12
+## 1.0.13
 
-- Added background color option(chat screen & chat listing)
+- added padding property in ChatScreen & ChatListScreen.
+- Added background color option(chat screen & chat listing).
 - Added background image option (chat screen).
 - added option to [disable/enable] delete option.
 - Fixed typingIndicationBuilder always showing issue.